add a notify/ subdirectory, move files accordingly
