+++ api.planx.uk/inviteToPay/sendPaymentEmail.ts
@@ -1,5 +1,6 @@
 import { gql } from "graphql-request";
-import { Template, calculateExpiryDate, convertSlugToName, getClientForTemplate, getHumanReadableProjectType, sendEmail } from "../saveAndReturn/utils";
+import { calculateExpiryDate, convertSlugToName, getHumanReadableProjectType } from "../saveAndReturn/utils";
+import { Template, getClientForTemplate, sendEmail } from "../notify/utils";
 import { InviteToPayNotifyConfig } from "../types";
 
 const sendSinglePaymentEmail = async (
