+++ api.planx.uk/notify/index.ts
@@ -1,0 +1,3 @@
+import { routeSendEmailRequest } from "./routeSendEmailRequest";
+
+export { routeSendEmailRequest };
