+++ api.planx.uk/saveAndReturn/index.ts
@@ -1,9 +1,7 @@
 import { resumeApplication } from "./resumeApplication";
 import { validateSession } from "./validateSession";
-import { routeSendEmailRequest } from "./sendEmail";
 
 export {
-  routeSendEmailRequest,
   resumeApplication,
   validateSession,
 };
