--- api.planx.uk/saveAndReturn/notify.ts
@@ -1,5 +1,0 @@
-import { NotifyClient } from "notifications-node-client";
-
-const notifyClient = new NotifyClient(process.env.GOVUK_NOTIFY_API_KEY);
-
-export { notifyClient };
